refactor: between

+++ src/between.cpp
@@ -1,4 +1,4 @@
-#include "between.hpp"
+#include "between.h"
 
 #include <random>
 
@@ -6,12 +6,12 @@ using std::mt19937;
 using std::random_device;
 using std::uniform_real_distribution;
 
-double between(const double start, const double end) {
+float between(const float start, const float end) {
   random_device device;
 
   mt19937 mersenne_twister(device());
 
-  uniform_real_distribution<double> distribution(start, end);
+  uniform_real_distribution<float> distribution(start, end);
 
   return distribution(mersenne_twister);
 }
